+++ examples/gradient.rs
@@ -5,7 +5,7 @@ use bevy::{
 use bevy_inspector_egui::WorldInspectorPlugin;
 use std::f32::consts::PI;
 
-use bevy_hanabi::*;
+use bevy_hanabi::prelude::*;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut options = WgpuSettings::default();